pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Alvin De Cruz <decruz@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
openai = "^0.27.4"
slack-bolt = "^1.17.1"
slack-sdk = "^3.21.1"
langchain = "^0.0.155"
pycoingecko = "^3.1.0"
redis = "^4.5.4"
Pillow = "^9.5.0"
stability-sdk = "^0.5.0"
wikipedia = "^1.4.0"
chromadb = "^0.3.21"
tiktoken = "^0.3.3"
playwright = "^1.32.1"
selenium = "^4.9.0"
unstructured = "^0.6.2"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"